set up embedded media package
