--- src/packages/core/modal/common/embedded-media/repository/index.ts
@@ -1,2 +1,0 @@
-export { UmbOEmbedRepository } from './oembed.repository.js';
-export { UMB_OEMBED_REPOSITORY_ALIAS } from './manifests.js';
