+++ src/packages/core/modal/common/index.ts
@@ -1,3 +1,2 @@
-export * from './embedded-media/index.js';
 export * from './confirm/index.js';
 export * from './item-picker/index.js';
