+++ src/packages/embedded-media/index.ts
@@ -1,1 +1,2 @@
+export * from './modal/index.js';
 export * from './repository/index.js';
