+++ src/packages/embedded-media/modal/embedded-media-modal.token.ts
