+++ src/packages/embedded-media/modal/index.ts
@@ -1,0 +1,1 @@
+export * from './embedded-media-modal.token.js';
