+++ src/packages/embedded-media/modal/manifests.ts
@@ -1,0 +1,8 @@
+export const manifests: Array<UmbExtensionManifest> = [
+	{
+		type: 'modal',
+		alias: 'Umb.Modal.EmbeddedMedia',
+		name: 'Embedded Media Modal',
+		element: () => import('./embedded-media-modal.element.js'),
+	},
+];
