+++ src/packages/embedded-media/repository/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_OEMBED_REPOSITORY_ALIAS = 'Umb.Repository.OEmbed';
