+++ src/packages/embedded-media/repository/index.ts
@@ -1,0 +1,2 @@
+export * from './oembed.repository.js';
+export * from './constants.js';
