+++ src/packages/embedded-media/umbraco-package.ts
@@ -1,0 +1,9 @@
+export const name = 'Umbraco.Core.EmbeddedMedia';
+export const extensions = [
+	{
+		name: 'Embedded Media Bundle',
+		alias: 'Umb.Bundle.EmbeddedMedia',
+		type: 'bundle',
+		js: () => import('./manifests.js'),
+	},
+];
